requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
IMAPClient==3.0.1
magic-filter==1.0.12
multidict==6.1.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
schedule==1.2.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36